requirements.txt
requests==2.31.0
pandas==2.0.3
openpyxl==3.1.2
pyzotero==1.5.19
fuzzywuzzy==0.18.0
python-levenshtein==0.21.1
scikit-learn==1.3.0
nltk==3.8.1
schedule==1.2.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.8.5
asyncio-mqtt==0.16.1
jupyter==1.0.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0